provide multi subpath exports support now this package can be imported as a pure module, or side effects only, or both.
