+++ package.json
@@ -1,14 +1,16 @@
 {
   "name": "@sec-ant/barcode-detector",
   "private": false,
-  "version": "1.0.2",
+  "version": "1.1.0",
   "type": "module",
   "files": [
     "./dist/**/*"
   ],
   "main": "./dist/index.js",
   "exports": {
-    ".": "./dist/index.js"
+    ".": "./dist/index.js",
+    "./pure": "./dist/pure.js",
+    "./side-effects": "./dist/side-effects.js"
   },
   "config": {
     "port": "18080"
