+++ src/index.ts
@@ -1,1 +1,2 @@
-export * from "./BarcodeDetector.js";
+import "./side-effects.js";
+export * from "./pure.js";
