+++ src/pure.ts
@@ -1,0 +1,1 @@
+export * from "./BarcodeDetector.js";
