+++ src/utils.ts
@@ -1,3 +1,20 @@
+export const BARCODE_DETECTOR_FORMATS = [
+  "aztec",
+  "code_128",
+  "code_39",
+  "code_93",
+  "codabar",
+  "data_matrix",
+  "ean_13",
+  "ean_8",
+  "itf",
+  "pdf417",
+  "qr_code",
+  "upc_a",
+  "upc_e",
+  "unknown",
+] as const;
+
 export function getIntrinsicDimensionsOfCanvasImageSource(
   image: CanvasImageSourceWebCodecs
 ): { width: number; height: number } {
